Address issue #383
- add checks for the local and remote time-axes being common earlier in the processing flow
- some factoring towards having better checks in future (towards using mth5's spectre.Spectrogram)
