+++ aurora/pipelines/time_series_helpers.py
@@ -8,8 +8,6 @@ import scipy.signal as ssig
 import xarray as xr
 
 from loguru import logger
-
-from aurora.time_series.windowed_time_series import WindowedTimeSeries
 from aurora.time_series.windowing_scheme import window_scheme_from_decimation
 from mt_metadata.transfer_functions.processing import TimeSeriesDecimation
 from mt_metadata.transfer_functions.processing.aurora.decimation_level import (
@@ -19,10 +17,7 @@ from mt_metadata.transfer_functions.processing.fourier_coefficients import (
     Decimation as FCDecimation,
 )
 from mth5.groups import RunGroup
-from mth5.processing.spectre.prewhitening import apply_prewhitening
-from mth5.processing.spectre.prewhitening import apply_recoloring
-import mth5.timeseries.spectre as spectre
-from typing import Literal, Optional, Union
+from typing import Union
 
 
 def truncate_to_clock_zero(
@@ -72,162 +67,6 @@ def truncate_to_clock_zero(
             logger.info(msg)
             run_xrds = run_xrds.where(cond1, drop=True)
     return run_xrds
-
-
-def nan_to_mean(xrds: xr.Dataset) -> xr.Dataset:
-    """
-    Set Nan values in xr.Dataset to the mean value per channel.
-
-    xrds: xr.Dataset
-        Time series data
-
-    Returns
-    -------
-    run_xrds : xr.Dataset
-        The same as the input time series but NaN values are replaced by the mean of the time series (per channel).
-
-    """
-    for ch in xrds.keys():
-        null_values_present = xrds[ch].isnull().any()
-        if null_values_present:
-            msg = "Unexpected Null values in xrds -- this should be examined"
-            logger.warning(msg)
-            value = np.nan_to_num(np.nanmean(xrds[ch].data))
-            xrds[ch] = xrds[ch].fillna(value)
-    return xrds
-
-
-def run_ts_to_stft(
-    decimation_obj: AuroraDecimationLevel, run_xrds_orig: xr.Dataset
-) -> spectre.Spectrogram:
-    """
-    Converts a runts object into a time series of Fourier coefficients.
-    Similar to run_ts_to_stft_scipy, but in this implementation operations on individual
-    windows are possible (for example pre-whitening per time window via ARMA filtering).
-
-    TODO: Make the output of this function a Spectrogram object
-
-    Parameters
-    ----------
-    decimation_obj : AuroraDecimationLevel
-        Information about how the decimation level is to be processed
-    run_xrds_orig: xarray.core.dataset.Dataset
-        normally extracted from mth5.RunTS
-
-    Returns
-    -------
-    stft_obj: xarray.core.dataset.Dataset
-        Note that the STFT object may have inf/nan in the DC harmonic, introduced by
-        recoloring. This really doesn't matter since we don't use the DC harmonic for
-        anything.
-    """
-    # need to remove any nans before windowing, or else if there is a single
-    # nan then the whole channel becomes nan.
-    run_xrds = nan_to_mean(run_xrds_orig)
-    run_xrds = apply_prewhitening(decimation_obj.stft.prewhitening_type, run_xrds)
-    run_xrds = truncate_to_clock_zero(decimation_obj, run_xrds)
-    windowing_scheme = window_scheme_from_decimation(decimation_obj)
-    windowed_obj = windowing_scheme.apply_sliding_window(
-        run_xrds, dt=1.0 / decimation_obj.decimation.sample_rate
-    )
-    if not np.prod(windowed_obj.to_array().data.shape):
-        raise ValueError
-
-    windowed_obj = WindowedTimeSeries.detrend(data=windowed_obj, detrend_type="linear")
-    tapered_obj = WindowedTimeSeries.apply_taper(
-        data=windowed_obj, taper=windowing_scheme.taper
-    )
-    stft_obj = WindowedTimeSeries.apply_fft(
-        data=tapered_obj,
-        sample_rate=windowing_scheme.sample_rate,
-        spectral_density_correction=windowing_scheme.linear_spectral_density_calibration_factor,
-        detrend_type=decimation_obj.stft.per_window_detrend_type,
-    )
-
-    if decimation_obj.stft.recoloring:
-        stft_obj = apply_recoloring(decimation_obj.stft.prewhitening_type, stft_obj)
-
-    spectrogram = spectre.Spectrogram(dataset=stft_obj)
-    return spectrogram
-
-
-def calibrate_stft_obj(
-    stft_obj: xr.Dataset,
-    run_obj: RunGroup,
-    units: Literal["MT", "SI"] = "MT",
-    channel_scale_factors: Optional[dict] = None,
-) -> xr.Dataset:
-    """
-    Calibrates frequency domain data into MT units.
-
-    Development Notes:
-     The calibration often raises a runtime warning due to DC term in calibration response = 0.
-     TODO: It would be nice to suppress this, maybe by only calibrating the non-dc terms and directly assigning np.nan to the dc component when DC-response is zero.
-
-    Parameters
-    ----------
-    stft_obj : xarray.core.dataset.Dataset
-        Time series of Fourier coefficients to be calibrated
-    run_obj : mth5.groups.master_station_run_channel.RunGroup
-        Provides information about filters for calibration
-    units : string
-        usually "MT", contemplating supporting "SI"
-    channel_scale_factors : Optional[dict]
-        Keyed by channel, supports a single scalar to apply to that channels data
-        Useful for debugging.  Should not be used in production and should throw a
-        warning if it is not None
-
-    Returns
-    -------
-    stft_obj : xarray.core.dataset.Dataset
-        Time series of calibrated Fourier coefficients
-    """
-    for channel_id in stft_obj.keys():
-
-        channel = run_obj.get_channel(channel_id)
-        channel_response = channel.channel_response
-        if not channel_response.filters_list:
-            msg = f"Channel {channel_id} with empty filters list detected"
-            logger.warning(msg)
-            if channel_id == "hy":
-                msg = "Channel hy has no filters, try using filters from hx"
-                logger.warning(msg)
-                channel_response = run_obj.get_channel("hx").channel_response
-
-        indices_to_flip = channel_response.get_indices_of_filters_to_remove(
-            include_decimation=False, include_delay=False
-        )
-        indices_to_flip = [
-            i for i in indices_to_flip if channel.metadata.filter.applied[i]
-        ]
-        filters_to_remove = [channel_response.filters_list[i] for i in indices_to_flip]
-        if not filters_to_remove:
-            logger.warning("No filters to remove")
-
-        calibration_response = channel_response.complex_response(
-            stft_obj.frequency.data, filters_list=filters_to_remove
-        )
-
-        if channel_scale_factors:
-            try:
-                channel_scale_factor = channel_scale_factors[channel_id]
-            except KeyError:
-                channel_scale_factor = 1.0
-            if channel_scale_factor != 1.0:
-                calibration_response /= channel_scale_factor
-
-        if units == "SI":
-            logger.warning("Warning: SI Units are not robustly supported issue #36")
-
-        # Handle case where DC term in calibration response = 0 (Sometimes raises a runtime warning)
-        zero_dc_term = calibration_response[0] == 0
-        if zero_dc_term:
-            calibration_response[0] = 1.0
-        stft_obj[channel_id].data /= calibration_response
-        if zero_dc_term:
-            stft_obj[channel_id].data[:, 0] = np.nan + 1j * np.nan
-
-    return stft_obj
 
 
 def prototype_decimate(
